fix: from float to int for price on publish menu

+++ Assets/1_FeedTheRealm/UI/MenuBar/FileOption/PublishMenu/PublishMenuController.cs
@@ -635,7 +635,7 @@ namespace FeedTheRealm.UI.MenuBar.FileOption.PublishMenu
             string categoryName,
             string fullPath,
             string existingSpriteId,
-            float price
+            int price
         )
         {
             if (!string.IsNullOrEmpty(existingSpriteId))
